Rename old gui to old config maker.
