+++ tools/old_config_maker/ninja_icon.gif
