+++ tools/old_config_maker/tab_manager.py
@@ -3,7 +3,7 @@ from tkinter.ttk import Notebook
 
 #from tools.config_maker.scroll_frame import VerticalScrolledFrame
 
-from tools.old_gui.graph_node import GraphNode
+from tools.old_config_maker.graph_node import GraphNode
 
 # Future
 # TODO -> should get data be put in graph node
